+++ bitwrk-blender/render_bitwrk/worker.py
@@ -1,7 +1,7 @@
 # ##### BEGIN GPL LICENSE BLOCK #####
 #
 #  BitWrk - A Bitcoin-friendly, anonymous marketplace for computing power
-#  Copyright (C) 2013-2017  Jonas Eschenburg <jonas@bitwrk.net>
+#  Copyright (C) 2013-2018  Jonas Eschenburg <jonas@bitwrk.net>
 #
 #  This program is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
@@ -59,6 +59,7 @@ def start_worker(settings):
         "--bitwrk-host", str(settings.bitwrk_client_host),
         "--bitwrk-port", str(settings.bitwrk_client_port),
         "--max-cost", str(settings.complexity),
+        "--device", str(settings.worker_device),
         ]
     print("Starting worker:", args)
     WORKER_PROC = subprocess.Popen(args)